Fix compilation on more modern linux systems
This tweaks a few things to make it compile out-of-the-box on modern
linux systems:

* Use PkgConfig to look up libtomcrypt if helper variables aren't set
* Explicitly include <cstring> and use std::memset and std::memcpy
* Detect the number of cores and parallelize the build

+++ psvpfsparser/FilesDbParser.cpp
@@ -449,7 +449,7 @@ bool constructFileMatrix(std::vector<sce_ng_pfs_block_t>& blocks, std::map<std::
             else
             {
                //empty files should be allowed!
-               std::cout << "[EMPTY] File " << fileName << " index " << child << " of type " << std::hex << block.m_infos[i].header.type << std::endl;
+               std::cout << "[EMPTY] File " << fileName << " index " << child << " of type " << std::hex << static_cast<std::uint16_t>(block.m_infos[i].header.type) << std::endl;
             }
          }
          else
